SmashExplorerCron/requirements.txt
azure-core==1.23.1
azure-cosmos==4.2.0
graphqlclient==0.2.4